rename editor widget to editor screen
