--- lib/router/edit/editor_widget.dart
@@ -1,499 +1,0 @@
-
-import 'package:bot_toast/bot_toast.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_quill/flutter_quill.dart';
-import 'package:intl/intl.dart';
-import 'package:nostrmo/nostr_sdk/nostr_sdk.dart';
-import 'package:nostrmo/component/editor/lnbc_embed_builder.dart';
-import 'package:nostrmo/component/editor/mention_event_embed_builder.dart';
-import 'package:nostrmo/component/editor/mention_user_embed_builder.dart';
-import 'package:nostrmo/component/editor/pic_embed_builder.dart';
-import 'package:nostrmo/component/editor/tag_embed_builder.dart';
-import 'package:nostrmo/component/editor/video_embed_builder.dart';
-import 'package:nostrmo/component/editor/zap_goal_input_widget.dart';
-import 'package:nostrmo/consts/base.dart';
-import 'package:nostrmo/main.dart';
-import 'package:nostrmo/router/index/index_app_bar.dart';
-import 'package:nostrmo/util/router_util.dart';
-
-import '../../component/appbar_back_btn_widget.dart';
-import '../../component/cust_state.dart';
-import '../../component/editor/custom_emoji_embed_builder.dart';
-import '../../component/editor/editor_mixin.dart';
-import '../../component/editor/poll_input_widget.dart';
-import '../../component/editor/zap_split_input_widget.dart';
-import '../../generated/l10n.dart';
-import 'editor_notify_item_widget.dart';
-
-class EditorWidget extends StatefulWidget {
-  static double appbarHeight = 56;
-
-  // dm arg
-  String? pubkey;
-
-  GroupIdentifier? groupIdentifier;
-
-  int? groupEventKind;
-
-  List<dynamic> tags = [];
-
-  List<dynamic> tagsAddedWhenSend = [];
-
-  List<dynamic> tagPs = [];
-
-  List<BlockEmbed>? initEmbeds;
-
-  bool isLongForm;
-
-  bool isPoll;
-
-  bool isZapGoal;
-
-  EditorWidget({
-    super.key,
-    required this.tags,
-    required this.tagsAddedWhenSend,
-    required this.tagPs,
-    this.pubkey,
-    this.initEmbeds,
-    this.groupIdentifier,
-    this.groupEventKind,
-    this.isLongForm = false,
-    this.isPoll = false,
-    this.isZapGoal = false,
-  });
-
-  static Future<Event?> open(
-    BuildContext context, {
-    List<dynamic>? tags,
-    List<dynamic>? tagsAddedWhenSend,
-    List<dynamic>? tagPs,
-    String? pubkey,
-    List<BlockEmbed>? initEmbeds,
-    GroupIdentifier? groupIdentifier,
-    int? groupEventKind,
-    bool isLongForm = false,
-    bool isPoll = false,
-    bool isZapGoal = false,
-  }) {
-    tags ??= [];
-    tagsAddedWhenSend ??= [];
-    tagPs ??= [];
-
-    var editor = EditorWidget(
-      tags: tags,
-      tagsAddedWhenSend: tagsAddedWhenSend,
-      tagPs: tagPs,
-      pubkey: pubkey,
-      initEmbeds: initEmbeds,
-      groupIdentifier: groupIdentifier,
-      groupEventKind: groupEventKind,
-      isLongForm: isLongForm,
-      isPoll: isPoll,
-      isZapGoal: isZapGoal,
-    );
-
-    return RouterUtil.push(context, MaterialPageRoute(builder: (context) {
-      return editor;
-    }));
-  }
-
-  @override
-  State<StatefulWidget> createState() {
-    return _EditorWidgetState();
-  }
-}
-
-class _EditorWidgetState extends CustState<EditorWidget> with EditorMixin {
-  List<EditorNotifyItem>? notifyItems;
-
-  List<EditorNotifyItem> editorNotifyItems = [];
-
-  @override
-  void initState() {
-    super.initState();
-    inputPoll = widget.isPoll;
-    inputZapGoal = widget.isZapGoal;
-    handleFocusInit();
-  }
-
-  @override
-  GroupIdentifier? getGroupIdentifier() {
-    return widget.groupIdentifier;
-  }
-
-  @override
-  int? getGroupEventKind() {
-    return widget.groupEventKind;
-  }
-
-  @override
-  bool isLongForm() {
-    return widget.isLongForm;
-  }
-
-  @override
-  Widget doBuild(BuildContext context) {
-    if (notifyItems == null) {
-      notifyItems = [];
-      for (var tagP in widget.tagPs) {
-        if (tagP is List<dynamic> && tagP.length > 1) {
-          notifyItems!.add(EditorNotifyItem(pubkey: tagP[1]));
-        }
-      }
-    }
-
-    final localization = S.of(context);
-    final themeData = Theme.of(context);
-    var hintColor = themeData.hintColor;
-    var textColor = themeData.textTheme.bodyMedium!.color;
-    var cardColor = themeData.cardColor;
-    var fontSize = themeData.textTheme.bodyMedium!.fontSize;
-    var largeTextSize = themeData.textTheme.bodyLarge!.fontSize;
-
-    List<Widget> list = [];
-
-    if (widget.tags.isNotEmpty) {
-      for (var tag in widget.tags) {
-        if (tag.length > 1) {
-          var tagName = tag[0];
-          var tagValue = tag[1];
-
-          if (tagName == "a") {
-            // this note is add to community
-            var aid = AId.fromString(tagValue);
-            if (aid != null && aid.kind == EventKind.COMMUNITY_DEFINITION) {
-              list.add(Container(
-                padding: const EdgeInsets.only(
-                  left: Base.basePadding,
-                  right: Base.basePadding,
-                ),
-                alignment: Alignment.centerLeft,
-                child: Row(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Container(
-                      margin: const EdgeInsets.only(right: Base.basePadding),
-                      child: Icon(
-                        Icons.groups,
-                        size: largeTextSize,
-                        color: hintColor,
-                      ),
-                    ),
-                    Text(
-                      aid.title,
-                      style: const TextStyle(
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ],
-                ),
-              ));
-            }
-          }
-        }
-      }
-    }
-
-    if ((notifyItems != null && notifyItems!.isNotEmpty) ||
-        (editorNotifyItems.isNotEmpty)) {
-      List<Widget> tagPsWidgets = [];
-      tagPsWidgets.add(Text("${localization.Notify}:"));
-      for (var item in notifyItems!) {
-        tagPsWidgets.add(EditorNotifyItemWidget(item: item));
-      }
-      for (var editorNotifyItem in editorNotifyItems) {
-        var exist = notifyItems!.any((element) {
-          return element.pubkey == editorNotifyItem.pubkey;
-        });
-        if (!exist) {
-          tagPsWidgets.add(EditorNotifyItemWidget(item: editorNotifyItem));
-        }
-      }
-      list.add(Container(
-        padding:
-            const EdgeInsets.only(left: Base.basePadding, right: Base.basePadding),
-        margin: const EdgeInsets.only(bottom: Base.basePaddingHalf),
-        width: double.maxFinite,
-        child: Wrap(
-          spacing: Base.basePaddingHalf,
-          runSpacing: Base.basePaddingHalf,
-          crossAxisAlignment: WrapCrossAlignment.center,
-          children: tagPsWidgets,
-        ),
-      ));
-    }
-
-    if (showTitle) {
-      list.add(buildTitleWidget());
-    }
-
-    if (isLongForm()) {
-      list.add(buildTitleWidget());
-
-      list.add(buildLongFormImageWidget());
-
-      list.add(buildSummaryWidget());
-    }
-
-    if (publishAt != null) {
-      var dateFormate = DateFormat("yyyy-MM-dd HH:mm");
-
-      list.add(GestureDetector(
-        onTap: selectedTime,
-        behavior: HitTestBehavior.translucent,
-        child: Container(
-          margin: const EdgeInsets.only(left: 10, bottom: Base.basePaddingHalf),
-          child: Row(
-            children: [
-              const Icon(Icons.timer_outlined),
-              Container(
-                margin: const EdgeInsets.only(left: 4),
-                child: Text(
-                  dateFormate.format(publishAt!),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ));
-    }
-
-    Widget quillWidget = QuillEditor(
-      configurations: QuillEditorConfigurations(
-        placeholder: localization.What_s_happening,
-        embedBuilders: [
-          MentionUserEmbedBuilder(),
-          MentionEventEmbedBuilder(),
-          PicEmbedBuilder(),
-          VideoEmbedBuilder(),
-          LnbcEmbedBuilder(),
-          TagEmbedBuilder(),
-          CustomEmojiEmbedBuilder(),
-        ],
-        scrollable: true,
-        autoFocus: false,
-        expands: false,
-        // padding: EdgeInsets.zero,
-        padding: const EdgeInsets.only(
-          left: Base.basePadding,
-          right: Base.basePadding,
-        ),
-        controller: editorController,
-      ),
-      scrollController: ScrollController(),
-      focusNode: focusNode,
-    );
-    List<Widget> editorList = [];
-    var editorInputWidget = Container(
-      margin: const EdgeInsets.only(bottom: Base.basePadding),
-      child: quillWidget,
-    );
-    editorList.add(editorInputWidget);
-    if (inputPoll) {
-      editorList.add(PollInputWidget(
-        pollInputController: pollInputController,
-      ));
-    }
-    if (inputZapGoal) {
-      editorList.add(ZapGoalInputWidget(
-        zapGoalInputController: zapGoalInputController,
-      ));
-    }
-    if (openZapSplit) {
-      editorList.add(ZapSplitInputWidget(eventZapInfos));
-    }
-
-    list.add(Expanded(
-      child: GestureDetector(
-        behavior: HitTestBehavior.translucent,
-        onTap: () {
-          // focus to editor input widget
-          focusNode.requestFocus();
-        },
-        child: Container(
-          constraints: BoxConstraints(
-              maxHeight: mediaDataCache.size.height -
-                  mediaDataCache.padding.top -
-                  EditorWidget.appbarHeight -
-                  IndexAppBar.height),
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: editorList,
-            ),
-          ),
-        ),
-      ),
-    ));
-
-    list.add(buildEditorBtns());
-    if (emojiShow) {
-      list.add(buildEmojiSelector());
-    }
-    if (customEmojiShow) {
-      list.add(buildEmojiListsWidget());
-    }
-
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: cardColor,
-        leading: const AppbarBackBtnWidget(),
-        actions: [
-          TextButton(
-            onPressed: documentSave,
-            style: const ButtonStyle(),
-            child: Text(
-              localization.Send,
-              style: TextStyle(
-                color: textColor,
-                fontSize: fontSize,
-              ),
-            ),
-          ),
-        ],
-      ),
-      body: Container(
-        color: cardColor,
-        child: Column(
-          children: list,
-        ),
-      ),
-    );
-  }
-
-  @override
-  Future<void> onReady(BuildContext context) async {
-    if (widget.initEmbeds != null && widget.initEmbeds!.isNotEmpty) {
-      {
-        final index = editorController.selection.baseOffset;
-        final length = editorController.selection.extentOffset - index;
-
-        editorController.replaceText(index, length, "\n", null);
-
-        editorController.moveCursorToPosition(index + 1);
-      }
-
-      for (var embed in widget.initEmbeds!) {
-        final index = editorController.selection.baseOffset;
-        final length = editorController.selection.extentOffset - index;
-
-        editorController.replaceText(index, length, embed, null);
-
-        editorController.moveCursorToPosition(index + 1);
-      }
-
-      editorController.moveCursorToPosition(0);
-    }
-
-    editorNotifyItems = [];
-    editorController.addListener(() {
-      bool updated = false;
-      Map<String, int> mentionUserMap = {};
-
-      var delta = editorController.document.toDelta();
-      var operations = delta.toList();
-      for (var operation in operations) {
-        if (operation.key == "insert") {
-          if (operation.data is Map) {
-            var m = operation.data as Map;
-            var value = m["mentionUser"];
-            if (StringUtil.isNotBlank(value)) {
-              mentionUserMap[value] = 1;
-            }
-          }
-        }
-      }
-
-      List<EditorNotifyItem> needDeleds = [];
-      for (var item in editorNotifyItems) {
-        var exist = mentionUserMap.remove(item.pubkey);
-        if (exist == null) {
-          updated = true;
-          needDeleds.add(item);
-        }
-      }
-      editorNotifyItems.removeWhere((element) => needDeleds.contains(element));
-
-      if (mentionUserMap.isNotEmpty) {
-        var entries = mentionUserMap.entries;
-        for (var entry in entries) {
-          updated = true;
-          editorNotifyItems.add(EditorNotifyItem(pubkey: entry.key));
-        }
-      }
-
-      if (updated) {
-        setState(() {});
-      }
-    });
-  }
-
-  Future<void> documentSave() async {
-    var cancelFunc = BotToast.showLoading();
-    try {
-      var event = await doDocumentSave();
-      if (event == null) {
-        BotToast.showText(text: S.of(context).Send_fail);
-        return;
-      }
-      RouterUtil.back(context, event);
-    } finally {
-      cancelFunc.call();
-    }
-  }
-
-  @override
-  BuildContext getContext() {
-    return context;
-  }
-
-  @override
-  void updateUI() {
-    setState(() {});
-  }
-
-  @override
-  String? getPubkey() {
-    return widget.pubkey;
-  }
-
-  @override
-  List getTags() {
-    return widget.tags;
-  }
-
-  @override
-  List getTagsAddedWhenSend() {
-    if ((notifyItems == null || notifyItems!.isEmpty) &&
-        editorNotifyItems.isEmpty) {
-      return widget.tagsAddedWhenSend;
-    }
-
-    List<dynamic> list = [];
-    list.addAll(widget.tagsAddedWhenSend);
-    for (var item in notifyItems!) {
-      if (item.selected) {
-        list.add(["p", item.pubkey]);
-      }
-    }
-
-    for (var editorNotifyItem in editorNotifyItems) {
-      var exist = notifyItems!.any((element) {
-        return element.pubkey == editorNotifyItem.pubkey;
-      });
-      if (!exist) {
-        if (editorNotifyItem.selected) {
-          list.add(["p", editorNotifyItem.pubkey]);
-        }
-      }
-    }
-
-    return list;
-  }
-
-  @override
-  bool isDM() {
-    return false;
-  }
-}
